feat: create routers structure, services layer / refactor routers to use services / update main.py to include routers
