+++ .env.example
@@ -1,5 +1,9 @@
+# параметры подключения к БД
 DB_HOST=localhost
 DB_PORT=5432
 DB_NAME=notes
 DB_USER=postgres
 DB_PASSWORD=postgres
+
+# строка подключения к БД
+DATABASE_URL=postgresql://db_user:db_password@db_host:db_port/db_name
