+++ backend/.dockerignore
@@ -10,8 +10,8 @@ __pycache__/
 .venv/
 
 # env
-.env
-!.env.example
+../.env
+!../.env.example
 
 # IDE
 .idea/
