+++ backend/app/models.py
@@ -3,10 +3,9 @@ from sqlalchemy.orm import relationship
 from .database import Base
 
 
-# параметр back_populates представляет атрибут связанной модели, с которой будет сопоставляться текущая модель
-# делаем связь Category.notes -- Note.category
+# Делаем связь Category.notes <-> Note.category
 
-# создаем модель категории, объекты которой будут храниться в бд
+# Модель категории, объекты которой будут храниться в БД
 class Category(Base):
     __tablename__ = "categories"
 
@@ -17,7 +16,7 @@ class Category(Base):
                          cascade="all, delete-orphan")  # связь с моделью Note через ее атрибут category
 
 
-# создаем модель заметки, объекты которой будут храниться в бд
+# Модель заметки, объекты которой будут храниться в БД
 class Note(Base):
     __tablename__ = "notes"
 
@@ -27,3 +26,6 @@ class Note(Base):
 
     category_id = Column(Integer, ForeignKey("categories.id"))  # внешний ключ на столбец id из таблицы "categories"
     category = relationship("Category", back_populates="notes")  # связь с моделью Category через ее атрибут notes
+
+# Примечание:
+# Параметр back_populates представляет атрибут связанной модели, с которой будет сопоставляться текущая модель
