+++ backend/requirements.txt
@@ -9,6 +9,7 @@ Markdown==3.10.2
 psycopg2-binary==2.9.11
 pydantic==2.12.5
 pydantic_core==2.41.5
+python-dotenv==1.2.2
 SQLAlchemy==2.0.48
 starlette==0.52.1
 typing-inspection==0.4.2
